Refactored main to multiple classes
