+++ observationer/lib/model/observation.dart
@@ -1,0 +1,11 @@
+class Observation {
+  Observation(this.id, this.subject, this.body, this.created, this.longitude,
+      this.latitude);
+
+  final int id;
+  final String subject;
+  final String body;
+  final String created;
+  final double longitude;
+  final double latitude;
+}
