+++ observationer/test/widget_test.dart
@@ -5,14 +5,9 @@
 // gestures. You can also use WidgetTester to find child widgets in the widget
 // tree, read text, and verify that the values of widget properties are correct.
 
-import 'package:flutter/material.dart';
-import 'package:flutter_test/flutter_test.dart';
-
-import 'package:observationer/main.dart';
-
-/**
- * Standard test implemented by Flutter-team below. Might be useful in the future.
- */
+///
+/// Standard test implemented by Flutter-team below. Might be useful in the future.
+///
 void main() {
   // testWidgets('Counter increments smoke test', (WidgetTester tester) async {
   //   // Build our app and trigger a frame.
